requirements.txt
Flask==0.10.1
Flask-RESTful==0.2.12
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.9.6
aniso8601==0.83
argparse==1.2.1
itsdangerous==0.24
pytz==2014.7
recaptcha-client==1.0.6
selenium==2.44.0
six==1.8.0
wsgiref==0.1.2
